+++ src/main/java/com/green/winey_final/admin/model/ProductUpdParam.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class ProductUpdParam {
-    private int productId;
+    private Long productId;
     private String nmKor;
     private String nmEng;
     private int price; //가격
@@ -16,15 +16,15 @@ public class ProductUpdParam {
     private int alcohol; //도수
     private int quantity; // 재고
 
-    private int country; //원산지
+    private Long country; //원산지
 
     private int sweety; //당도
     private int acidity; //산도
     private int body; //바디
 
-    private int category; //레드(1) 화이트(2) 스파클링(3) 기타(4) //
+    private Long category; //레드(1) 화이트(2) 스파클링(3) 기타(4) //
 
-    private List<Integer> aroma; //향
+    private List<Long> aroma; //향
 
     private int sale; //할인률
     private int salePrice; //할인가격
