+++ src/main/java/com/green/winey_final/repository/SmallCategoryRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SmallCategoryRepository extends JpaRepository<SmallCategoryEntity, Long> {}
+public interface SmallCategoryRepository extends JpaRepository<SmallCategoryEntity, Integer> {}
